Code optimizations

+++ src/pipeline_base.cpp
@@ -121,7 +121,7 @@ void on_message_element(const GstBus* gst_bus, GstMessage* message, PipelineBase
 
     magnitudes = gst_structure_get_value(s, "magnitude");
 
-    for (uint n = 0; n < pb->spectrum_freqs.size(); n++) {
+    for (uint n = 0u; n < pb->spectrum_freqs.size(); n++) {
       pb->spectrum_mag_tmp[n] = g_value_get_float(gst_value_list_get_value(magnitudes, n + pb->spectrum_start_index));
     }
 
@@ -129,7 +129,7 @@ void on_message_element(const GstBus* gst_bus, GstMessage* message, PipelineBase
       boost::math::interpolators::cardinal_cubic_b_spline<float> spline(
           pb->spectrum_mag_tmp.begin(), pb->spectrum_mag_tmp.end(), pb->spline_f0, pb->spline_df);
 
-      for (uint n = 0; n < pb->spectrum_mag.size(); n++) {
+      for (uint n = 0u; n < pb->spectrum_mag.size(); n++) {
         pb->spectrum_mag[n] = spline(pb->spectrum_x_axis[n]);
       }
     } catch (const std::exception& e) {
@@ -154,14 +154,15 @@ void on_message_element(const GstBus* gst_bus, GstMessage* message, PipelineBase
 }
 
 void on_spectrum_n_points_changed(GSettings* settings, gchar* key, PipelineBase* pb) {
-  long unsigned int npoints = g_settings_get_int(settings, "n-points");
+  unsigned long int npoints = g_settings_get_int(settings, "n-points");
 
   if (npoints != pb->spectrum_mag.size()) {
     pb->resizing_spectrum = true;
 
     pb->spectrum_mag.resize(npoints);
 
-    pb->spectrum_x_axis = util::logspace(log10(pb->min_spectrum_freq), log10(pb->max_spectrum_freq), npoints);
+    pb->spectrum_x_axis = util::logspace(log10(static_cast<float>(pb->min_spectrum_freq)),
+                                         log10(static_cast<float>(pb->max_spectrum_freq)), npoints);
 
     pb->resizing_spectrum = false;
   }
@@ -662,9 +663,9 @@ void PipelineBase::init_spectrum() {
 
   spectrum_freqs.clear();
 
-  spectrum_start_index = 0;
+  spectrum_start_index = 0u;
 
-  for (uint n = 0; n < spectrum_nbands; n++) {
+  for (uint n = 0u; n < spectrum_nbands; n++) {
     auto f = sampling_rate * (0.5 * n + 0.25) / spectrum_nbands;
 
     if (f > max_spectrum_freq) {
@@ -674,7 +675,7 @@ void PipelineBase::init_spectrum() {
     if (f > min_spectrum_freq) {
       spectrum_freqs.emplace_back(f);
 
-      if (spectrum_start_index == 0) {
+      if (spectrum_start_index == 0u) {
         spectrum_start_index = n;
       }
     }
@@ -685,7 +686,8 @@ void PipelineBase::init_spectrum() {
 
     auto npoints = g_settings_get_int(spectrum_settings, "n-points");
 
-    spectrum_x_axis = util::logspace(log10(min_spectrum_freq), log10(max_spectrum_freq), npoints);
+    spectrum_x_axis = util::logspace(log10(static_cast<float>(min_spectrum_freq)),
+                                     log10(static_cast<float>(max_spectrum_freq)), npoints);
 
     spectrum_mag.resize(npoints);
 
@@ -695,9 +697,9 @@ void PipelineBase::init_spectrum() {
 }
 
 void PipelineBase::update_spectrum_interval(const double& value) const {
-  const float one_second_in_ns = 1000000000.0F;
+  const double one_second_in_ns = 1000000000.0;
 
-  auto interval = guint64(one_second_in_ns / value);
+  auto interval = static_cast<guint64>(one_second_in_ns / value);
 
   g_object_set(spectrum, "interval", interval, nullptr);
 }
